combined reader and writer class
