+++ src/libxtp/molden.cc
@@ -1,14 +1,100 @@
 #include <sstream>
 
-#include "votca/xtp/moldenreader.h"
+#include "votca/xtp/basisset.h"
+#include "votca/xtp/molden.h"
+#include <boost/algorithm/string.hpp>
 
 namespace votca {
 namespace xtp {
 
-// The returned string contains the line with the next section header
-// or if it is the last section an empty string.
-std::string MoldenReader::readAtoms(QMMolecule& mol, const std::string& units,
-                                    std::ifstream& input_file) const {
+void Molden::writeAtoms(const Orbitals& orbitals,
+                        std::ofstream& outFile) const {
+  for (const auto& atom : orbitals.QMAtoms()) {
+    const Eigen::Vector3d& pos = atom.getPos();
+    outFile << boost::format("%-4s %5d %5d %22.12e %22.10e %22.10e\n") %
+                   atom.getElement() % (atom.getId() + 1) %
+                   atom.getTotalNuccharge() % pos[0] % pos[1] % pos[2];
+  }
+}
+
+void Molden::writeMOs(const Orbitals& orbitals, std::ofstream& outFile) const {
+  Eigen::VectorXd energies = orbitals.MOs().eigenvalues();
+  OrbReorder reorder(_reorderList, _multipliers);
+  Eigen::MatrixXd moCoefficients = orbitals.MOs().eigenvectors();
+  reorder.reorderOrbitals(moCoefficients, _basis);
+
+  for (Index i = 0; i < orbitals.getBasisSetSize(); i++) {  // over columns
+    outFile << "Sym= \n";
+    outFile << boost::format("Ene= %-20.12e\n") % energies[i];
+    outFile << "Spin= Alpha\n";
+    outFile << boost::format("Occup= %-5.2f\n") %
+                   (2 * (i < orbitals.getLumo()));
+    for (Index j = 0; j < orbitals.getBasisSetSize(); j++) {
+      outFile << boost::format("%5d %22.12e\n") % (j + 1) %
+                     moCoefficients(j, i);
+    }
+  }
+}
+
+void Molden::writeBasisSet(const Orbitals& orbitals,
+                           std::ofstream& outFile) const {
+
+  for (const auto& atom : orbitals.QMAtoms()) {
+    const Element& element = _bs.getElement(atom.getElement());
+    // The 0 in the format string of the next line is meaningless it
+    // is included for backwards compatibility of molden files
+    outFile << boost::format("%4d 0 \n") % (atom.getId() + 1);
+    for (const Shell& shell : element) {
+      // The 1.0 at the end of the next line is meaningless it
+      // is included for backwards compatibility of molden files
+      outFile << boost::format("%-3s %4d %3.1f \n") %
+                     boost::to_lower_copy(EnumToString(shell.getL())) %
+                     shell.getSize() % 1.0;
+      for (const GaussianPrimitive& gaussian : shell) {
+        outFile << boost::format("%22.10e %22.10e\n") % gaussian.decay() %
+                       gaussian.contraction();
+      }
+    }
+    outFile << " \n";
+  }
+}
+
+void Molden::WriteFile(const std::string& filename, const Orbitals& orbitals) {
+  if (orbitals.hasDFTbasisName()) {
+    _bs.Load(orbitals.getDFTbasisName());
+    _basis.Fill(_bs, orbitals.QMAtoms());
+  } else {
+    throw std::runtime_error(".orb file does not contain a basisset name");
+  }
+
+  std::ofstream outFile(filename);
+  if (outFile.is_open()) {
+
+    XTP_LOG(Log::error, _log) << "Writing data to " << filename << std::flush;
+
+    // print Header
+    outFile << "[Molden Format]\n";
+    outFile << "[Title]\n";
+    outFile << "Molden file created by VOTCA-XTP\n";
+    outFile << " \n";
+
+    outFile << "[Atoms] AU\n";
+    writeAtoms(orbitals, outFile);
+
+    outFile << "[GTO] \n";
+    writeBasisSet(orbitals, outFile);
+
+    // indicate spherical D F and G functions
+    outFile << "[5D] \n[7F] \n[9G] \n";
+
+    outFile << "[MO]\n";
+    writeMOs(orbitals, outFile);
+    XTP_LOG(Log::error, _log)
+        << "Finished writing to molden file." << std::flush;
+  }
+}
+std::string Molden::readAtoms(QMMolecule& mol, const std::string& units,
+                              std::ifstream& input_file) const {
   std::string line;
   std::istringstream iss(" ");
   while (std::getline(input_file, line)) {
@@ -40,8 +126,8 @@ std::string MoldenReader::readAtoms(QMMolecule& mol, const std::string& units,
 
 // The returned string contains the line with the next section header
 // or, if it is the last section, an empty string.
-std::string MoldenReader::readMOs(Orbitals& orbitals,
-                                  std::ifstream& input_file) const {
+std::string Molden::readMOs(Orbitals& orbitals,
+                            std::ifstream& input_file) const {
 
   // setup space to store everything
   Index basis_size = orbitals.getBasisSetSize();
@@ -103,7 +189,7 @@ std::string MoldenReader::readMOs(Orbitals& orbitals,
   return line;
 }
 
-void MoldenReader::addBasissetInfo(Orbitals& orbitals) {
+void Molden::addBasissetInfo(Orbitals& orbitals) {
   BasisSet bs;
   bs.Load(_basisset_name);
 
@@ -113,8 +199,7 @@ void MoldenReader::addBasissetInfo(Orbitals& orbitals) {
   orbitals.setAuxbasisName(_aux_basisset_name);
 }
 
-void MoldenReader::parseMoldenFile(const std::string& filename,
-                                   Orbitals& orbitals) {
+void Molden::parseMoldenFile(const std::string& filename, Orbitals& orbitals) {
 
   if (_basisset_name == "") {
     throw std::runtime_error(
