+++ src/tests/CMakeLists.txt
@@ -75,8 +75,7 @@ if(ENABLE_TESTING)
   list(APPEND test_cases test_indexparser)
   list(APPEND test_cases test_settings)
   list(APPEND test_cases test_orbreorder)
-  list(APPEND test_cases test_moldenreader)
-  list(APPEND test_cases test_moldenwriter)
+  list(APPEND test_cases test_molden)
   if(USE_CUDA)
     list(APPEND test_cases test_cudapipeline)
     list(APPEND test_cases test_cudamatrix)
