+++ src/tests/test_molden.cc
@@ -15,7 +15,7 @@
  */
 #define BOOST_TEST_MAIN
 
-#define BOOST_TEST_MODULE moldenwriter_test
+#define BOOST_TEST_MODULE moldenreader_test
 
 // Third party includes
 #include <boost/test/unit_test.hpp>
@@ -26,15 +26,44 @@
 
 // Local VOTCA includes
 #include "votca/xtp/logger.h"
-#include "votca/xtp/moldenreader.h"
-#include "votca/xtp/moldenwriter.h"
+#include "votca/xtp/molden.h"
 #include "votca/xtp/orbitals.h"
 
 using namespace votca::xtp;
 using namespace votca;
 using namespace std;
 
-BOOST_AUTO_TEST_SUITE(moldenwriter_test)
+BOOST_AUTO_TEST_SUITE(molden_test)
+
+BOOST_AUTO_TEST_CASE(moldenreader_test) {
+
+  Eigen::MatrixXd coeffs_ref = votca::tools::EigenIO_MatrixMarket::ReadMatrix(
+      std::string(XTP_TEST_DATA_FOLDER) + "/molden/orbitalsMOs_ref.mm");
+
+  Orbitals orbitals_ref;
+  orbitals_ref.QMAtoms().LoadFromFile(std::string(XTP_TEST_DATA_FOLDER) +
+                                      "/molden/benzene.xyz");
+
+  Logger log;
+  Molden molden(log);
+  molden.setBasissetInfo(
+      std::string(XTP_TEST_DATA_FOLDER) + "/molden/def2-tzvp.xml",
+      "aux-def2-tzvp");
+  Orbitals orbitals;
+  molden.parseMoldenFile(
+      std::string(XTP_TEST_DATA_FOLDER) + "/molden/benzene.molden.input",
+      orbitals);
+
+  // Check if MO's are read correctly
+  BOOST_CHECK(orbitals.MOs().eigenvectors().isApprox(coeffs_ref, 1e-5));
+
+  // Check if atoms are read correctly
+  BOOST_CHECK(orbitals.QMAtoms().size() == orbitals_ref.QMAtoms().size());
+  for (int i = 0; i < orbitals.QMAtoms().size(); i++) {
+    BOOST_CHECK(orbitals.QMAtoms()[i].getPos().isApprox(
+        orbitals_ref.QMAtoms()[i].getPos(), 1e-3));
+  }
+}
 
 BOOST_AUTO_TEST_CASE(moldenwriter_test) {
 
@@ -85,11 +114,10 @@ BOOST_AUTO_TEST_CASE(moldenwriter_test) {
 
   // write orbitals object to molden file
   Logger log;
-  MoldenWriter moldenWriter(log);
-  moldenWriter.WriteFile("moldenFile.molden", orbitals_ref);
+  Molden molden(log);
+  molden.WriteFile("moldenFile.molden", orbitals_ref);
 
   // read in written molden file
-  MoldenReader molden(log);
   molden.setBasissetInfo(
       std::string(XTP_TEST_DATA_FOLDER) + "/molden/def2-tzvp.xml",
       "aux-def2-tzvp");
